fix: use long sum for resource count in text logger
The Sum of ResourceCount values overflowed Int32 when total resource
usage exceeded ~2.1 billion. Cast to long before summing.

Fixes #6281

+++ Source/DafnyDriver/JsonVerificationLogger.cs
@@ -89,7 +89,7 @@ public class JsonVerificationLogger : IVerificationResultFormatLogger {
       ["name"] = scope.Name,
       ["outcome"] = SerializeOutcome(results.Aggregate(VcOutcome.Correct, (o, r) => MergeOutcomes(o, r.Result.Outcome))),
       ["runTime"] = SerializeTimeSpan(TimeSpan.FromSeconds(results.Sum(r => r.Result.RunTime.Seconds))),
-      ["resourceCount"] = results.Sum(r => r.Result.ResourceCount),
+      ["resourceCount"] = results.Sum(r => (long)r.Result.ResourceCount),
       ["vcResults"] = new JsonArray(results.Select(r => SerializeVcResult(dependencyManager, potentialDependencies, r)).ToArray())
     };
     if (potentialDependencies is not null) {
